+ <le:manager> @call - window\\Container::initMessages() used by binder\\JSON and html\\Document + modules\html\\backend.css .loading, loader.gif - modules\html : form .actions > form_actions.less + modules\less\Prefixer column-count, column-fill + <tst:test> @run + modules/todo/explorer (wip) - modules/todo/schema.xql updated with duration, term, owner, delegate, parent and tags - .form-actions > .actions - modules/uploader/fieldset.tpl .actions > .item-actions + languages\common\\Window::createWhile(), + php\\_While, <php:while> + php\\Window::createKey(), + php\\_VarKey, <php:key> - php\\_Switch::setContents() can create case + schema\cached\view\_String /format(end) + sql/crud.crd + <sql:foreign> @key, sql\schema\\Foreign::load,set,getKey() + sql\template\\CollectionGroup with /group() + sql\template\pathable::reflectRegister() - sql\test\samples\user1.xql fixed bug xsd string not working - sql\update.crd fixed bug bad xmode - WARNING : causes foreign error, to fix - template\parser\Pather::parseExpressionValue() read path instead of apply - <tpl:else> fixed bug no parse + <tpl:switch> + ui.Loader.options.loaderPosition - ui.Main.loadArray() fixed bug firing load events

+++ trunk/modules/tester/Parser.php
@@ -93,7 +93,10 @@ class Parser extends Prepare {
         }
       }
 
-      $result = $this->loadResult($manager, $file, $aArguments, !$test->getx('self:expected', array(), false));
+      $bExpected = !$test->getx('self:expected', array(), false);
+      $sRun = $test->readx('@run', array(), false);
+
+      $result = $this->loadResult($manager, $file, $aArguments, $sRun !== 'false', $bExpected);
 
     } catch (\sylma\core\exception $e) {
 
@@ -108,12 +111,16 @@ class Parser extends Prepare {
     return $this->catchExceptionCheck($test->readAttribute(self::PARSER_EXCEPTION, null, false), $test, $e, $file);
   }
 
-  protected function loadResult($manager, fs\file $file, array $aArguments, $bDelete = true) {
+  protected function loadResult($manager, fs\file $file, array $aArguments, $bRun = true, $bDelete = true) {
 
+    $result = null;
     $this->setFile($file);
 
-    $result = $manager->load($file, $aArguments, false, static::DEBUG_RUN);
-    if ($bDelete) $file->delete();
+    if ($bRun) {
+
+      $result = $manager->load($file, $aArguments, false, static::DEBUG_RUN);
+      if ($bDelete) $file->delete();
+    }
 
     return $result;
   }
